Isolate most of the softgpu specialization code to RasterizerRectangle.
See comments.
