+++ GPU/GPU.vcxproj.filters
@@ -282,6 +282,9 @@
     <ClInclude Include="Debugger\RecordFormat.h">
       <Filter>Debugger</Filter>
     </ClInclude>
+    <ClInclude Include="Software\RasterizerRectangle.h">
+      <Filter>Software</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="Math3D.cpp">
@@ -559,6 +562,9 @@
     </ClCompile>
     <ClCompile Include="Debugger\Playback.cpp">
       <Filter>Debugger</Filter>
+    </ClCompile>
+    <ClCompile Include="Software\RasterizerRectangle.cpp">
+      <Filter>Software</Filter>
     </ClCompile>
   </ItemGroup>
 </Project>
