+++ GPU/Software/Clipper.cpp
@@ -17,19 +17,13 @@
 
 #include <algorithm>
 
-#include "Core/System.h"
-
 #include "GPU/GPUState.h"
 
 #include "GPU/Software/Clipper.h"
 #include "GPU/Software/Rasterizer.h"
+#include "GPU/Software/RasterizerRectangle.h"
 
 #include "profiler/profiler.h"
-
-
-extern bool g_DarkStalkerStretch;
-// For Darkstalkers hack. Ugh.
-extern bool currentDialogActive;
 
 namespace Clipper {
 
@@ -141,11 +135,8 @@ static void RotateUVThrough(const VertexData &tl, const VertexData &br, VertexDa
 	}
 }
 
-bool needsClear = false;
-
 void ProcessRect(const VertexData& v0, const VertexData& v1)
 {
-	g_DarkStalkerStretch = false;
 	if (!gstate.isModeThrough()) {
 		VertexData buf[4];
 		buf[0].clippos = ClipCoords(v0.clippos.x, v0.clippos.y, v1.clippos.z, v1.clippos.w);
@@ -188,41 +179,8 @@ void ProcessRect(const VertexData& v0, const VertexData& v1)
 	} else {
 		// through mode handling
 
-		// Check for 1:1 texture mapping. In that case we can call DrawSprite.
-		int xdiff = v1.screenpos.x - v0.screenpos.x;
-		int ydiff = v1.screenpos.y - v0.screenpos.y;
-		int udiff = (v1.texturecoords.x - v0.texturecoords.x) * 16.0f;
-		int vdiff = (v1.texturecoords.y - v0.texturecoords.y) * 16.0f;
-		bool coord_check =
-			(xdiff == udiff || xdiff == -udiff) &&
-			(ydiff == vdiff || ydiff == -vdiff);
-		bool state_check = !gstate.isModeClear();  // TODO: Add support for clear modes in Rasterizer::DrawSprite.
-		if ((coord_check || !gstate.isTextureMapEnabled()) && state_check) {
-			Rasterizer::DrawSprite(v0, v1);
+		if (Rasterizer::RectangleFastPath(v0, v1)) {
 			return;
-		}
-
-		// Eliminate the stretch blit in DarkStalkers.
-		// We compensate for that when blitting the framebuffer in SoftGpu.cpp.
-		if (PSP_CoreParameter().compat.flags().DarkStalkersPresentHack && v0.texturecoords.x == 64.0f && v0.texturecoords.y == 16.0f && v1.texturecoords.x == 448.0f && v1.texturecoords.y == 240.0f) {
-			if (v0.screenpos.x == 0x7100 && v0.screenpos.y == 0x7780 && v1.screenpos.x == 0x8f00 && v1.screenpos.y == 0x8880) {
-				// Also check for save/load dialog.
-				if (!currentDialogActive) {
-					g_DarkStalkerStretch = true;
-					if (needsClear) {
-						needsClear = false;
-						// Afterwards, we also need to clear the actual destination. Can do a fast rectfill.
-						gstate.textureMapEnable &= ~1;
-						VertexData newV0 = v0;
-						newV0.color0 = Vec4<int>(0, 0, 0, 255);
-						Rasterizer::DrawSprite(newV0, v1);
-						gstate.textureMapEnable |= 1;
-					}
-					return;
-				} else {
-					needsClear = true;
-				}
-			} // else, handle the Capcom screen stretch, or the non-wide stretch? Or let's just not bother.
 		}
 
 		VertexData buf[4];
