+++ GPU/Software/Rasterizer.h
@@ -27,10 +27,14 @@ namespace Rasterizer {
 void DrawTriangle(const VertexData& v0, const VertexData& v1, const VertexData& v2);
 void DrawPoint(const VertexData &v0);
 void DrawLine(const VertexData &v0, const VertexData &v1);
-void DrawSprite(const VertexData &v0, const VertexData &v1);
 void ClearRectangle(const VertexData &v0, const VertexData &v1);
 
 bool GetCurrentStencilbuffer(GPUDebugBuffer &buffer);
 bool GetCurrentTexture(GPUDebugBuffer &buffer, int level);
 
-}
+// Shared functions with RasterizerRectangle.cpp
+Vec3<int> AlphaBlendingResult(const Vec4<int> &source, const Vec4<int> &dst);
+void DrawSinglePixelNonClear(const DrawingCoords &p, u16 z, u8 fog, const Vec4<int> &color_in);
+Vec4<int> GetTextureFunctionOutput(const Vec4<int>& prim_color, const Vec4<int>& texcolor);
+
+}  // namespace Rasterizer
